pls
ugh

+++ code/game/machinery/cryopod.dm
@@ -307,6 +307,7 @@
 	//Handle Borg stuff first
 	if(iscyborg(mob_occupant))
 		var/mob/living/silicon/robot/R = mob_occupant
+		if(!istype(R)) return ..()
 
 		R.contents -= R.mmi
 		qdel(R.mmi)
@@ -369,14 +370,16 @@
 					O.find_target()
 					O.update_explanation_text()
 					if(!(O.target))
+						O.owner.objectives -= O
 						qdel(O)
 
-	if(mob_occupant.mind)
+	if(mob_occupant.mind && mob_occupant.mind.assigned_role)
 		//Handle job slot/tater cleanup.
-		if(mob_occupant.mind.assigned_role)
-			var/job = mob_occupant.mind.assigned_role
-			SSjob.FreeRole(job)
-		mob_occupant.mind.special_role = null
+		var/job = mob_occupant.mind.assigned_role
+		SSjob.FreeRole(job)
+		if(mob_occupant.mind.objectives.len)
+			mob_occupant.mind.objectives.Cut()
+			mob_occupant.mind.special_role = null
 
 	// Delete them from datacore.
 
@@ -408,7 +411,7 @@
 
 	// Ghost and delete the mob.
 	if(!mob_occupant.get_ghost(1))
-		mob_occupant.ghostize(FALSE, penalize = TRUE)
+		mob_occupant.ghostize(0) // Players who cryo out may not re-enter the round
 
 	QDEL_NULL(occupant)
 	open_machine()
